Eclipse3.6M6 vA39 patched JDT Core
git-svn-id: https://svn.codehaus.org/groovy/eclipse/trunk@19841 a5544e8c-8a19-0410-ba12-f9af4593a198

+++ base/org.eclipse.jdt.core/model/org/eclipse/jdt/internal/core/JavaModelCache.java
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2000, 2008 IBM Corporation and others.
+ * Copyright (c) 2000, 2010 IBM Corporation and others.
  * All rights reserved. This program and the accompanying materials
  * are made available under the terms of the Eclipse Public License v1.0
  * which accompanies this distribution, and is available at
@@ -155,7 +155,7 @@ public IJavaElement getExistingElement(IJavaElement element) {
 }
 
 protected double getMemoryRatio() {
-	if (this.memoryRatio == -1) {
+	if ((int) this.memoryRatio == -1) {
 		long maxMemory = Runtime.getRuntime().maxMemory();
 		// if max memory is infinite, set the ratio to 4d which corresponds to the 256MB that Eclipse defaults to
 		// (see https://bugs.eclipse.org/bugs/show_bug.cgi?id=111299)
